Rename question specific files
